Reduce the needed memory for compilation
The linking of rootcling_stage1 and libCling requires a lot of memory.
Since these are linked from mostly the same objects, the build is
ready to link them at the same time. If you make a parallel build this
means that the two targets that require the most amount of memory are
being linked in parallel. This exhausts the available memory, and the
computer starts swapping.

This commit adds a dependency of one of the targets to the other. The
dependency is not really there since it is not needed for building,
but it prevents the two memory consuming targets to be built in
parallel.

A similar dependency existed before the latest code changes (see
commit 2638f6fc7f54b0995f2f9d60363daaf8aae2386e), then between
rootcling and libCling.